requirements.txt
Flask
Flask-SQLAlchemy
reportlab
pytesseract
pdf2image
PyPDF2
Pillow
